ref(extension): Split content and options entrypoints into focused modules
Move the remaining hot-path logic out of the oversized content and
options entrypoints so src/content/bootstrap.js and src/options/index.js
become orchestration layers instead of carrying extraction, overlay,
history, and settings internals directly. This keeps behavior locked by
existing release verification while lowering the cost of future changes.

Constraint: Behavior must continue to match the frozen baseline commit c911c73
Rejected: Leave bootstrap and options index as monoliths | future changes would keep compounding maintenance cost
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Continue new work in the extracted src/content and src/options modules rather than rebuilding logic in the entrypoints
Tested: npm run verify:release
Not-tested: Extra exploratory scenarios outside the automated smoke suite
